requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
google-generativeai>=0.3.2
Pillow>=10.1.0
Jinja2>=3.1.2
python-magic>=0.4.27
httpx>=0.25.0
aiohttp>=3.9.0
python-multipart>=0.0.6
uv